Iteratively process basic blocks in try bodies
Summary:
We currently recursively DFS through the CFG when identifying basic
blocks associated with each catch. For very deep CFGs this can cause us
to exceed the max stack depth and incorrectly report an error about
nested try/catch statements.

This behaviour is easily triggered if we have a huge number of
try/catch blocks in a given function, because we then have 3 stack
frames for each try (one each for the entry, body, and end).

This diff rewrites the traversal code so that we process the entire try
body iteratively, and only use recursion when we encounter another try
within a try body.

Reviewed By: tmikov

Differential Revision: D29259502

fbshipit-source-id: 5867847094ebf83d539c6774876e0bfefe229d06