Delete ANONYMOUS_PROPERTY_SLOTS mechanism
Summary:
There are two advantages of the `ANONYMOUS_PROPERTY_SLOTS`.
1. They maintain the metadata for marking automatically.
2. They can be used to work around alignment constraints with overlap
slots. For example if some derived class is 8 byte aligned, and is
actually 4 bytes smaller than `JSObjectWithDirectProps`, the last
direct property slot is lost to alignment. In such a case, we can use
an anonymous property slot store one of the derived class'
properties, and reclaim that last 4 bytes.

However, they also add significant complexity, since we have to keep
track of the exact index where a property is going to be and how many
of the direct slots are used for this mechanism. They can also be
relatively inefficient, since most internal properties have a known
type, and it would be faster to use a `GCPointer` or something else
that more directly represents the type.

The point on alignment is also less of a concern now, because GCCell
headers are now 4-bytes, which means that JSObject itself is 4-byte
aligned. Derived classes with 8 byte members may still encounter this
problem though (e.g. something that stores a double, native pointer,
or HV64).

Reviewed By: kodafb

Differential Revision: D28465394

fbshipit-source-id: fa15646eca1600bc25061e146fbdecef3ed4ea55