Handle starting OG collection with pending compactee
Summary:
We depend on having all parts of an existing OG collection completed
before starting a new one. For compactions, we can typically depend on
the fact that if the OG collection state is `None`, and we are at the
end of a YG collection, then the current YG collection must have also
completed any pending compaction.

However, this does not deal with the case where the current YG
collection had to block on the completion of an OG collection. In that
case, the phase could be `None` without the current `compactee_` being
evacuated.

To fix this, check whether there is a compactee ready to be evacuated
even when the phase is `None`.

Reviewed By: kodafb

Differential Revision: D29438014

fbshipit-source-id: b43233e7c3751fb17bbfc07724f7aa66ae0e80ec