Delete allowFunctionToStringWithRuntimeSource
Summary:
This diff deprecate an old mechanism to allow `Function$toString`
to use source code if the JS was compiled at runtime (eval, lazy),
in which case the source code is available at runtime. However,
this mechanism cannot be scaled to JS functions pre-compiled
to HBC.

In later diff, we'll introduce an new mechanism which reserves
function source through HBC so it's general enough to be working
for both the eager and lazy compilation.

Reviewed By: avp

Differential Revision: D29093332

fbshipit-source-id: 932b4c588f521bd13adba43f34347d255529bc70