Fix ssize_t conflict in 32 bit Windows
Summary:
We currently have a conflict between zip.h and LLVM's DataTypes.h since
both defined `ssize_t` on Windows.

This is happening because LLVM typedefs it to `int` on 32 bit whereas
zip typedefs it to `long`. Since MSVC itself typedefs `intptr_t` to
`int` and not `long` on 32 bit, this diff patches zip to use
`intptr_t`, which resolves the conflict by making the typedefs
equivalent.

Reviewed By: kodafb

Differential Revision: D29601966

fbshipit-source-id: 28da7491aa3a0f8f4926caa32885f6ed5ea2efd8