Refactor how directives e.g. strictness are handled for lazy
Summary:
The current approach of handling directives limits what we can do,
and the logic of computing strictness is duplicated and confusing.

The goal of this diff is to first achieve parity with the previous
strictness behaviors, and then migrate/implement other/new
directives to this new approach in upper diffs.

### Status Quo

Parser is reponsible to set strictness to ESTree Node
1. preParse records strictness for every function
2. lazyParse recovers info from preParse to maintain the invariant
that parser can still set the strictness correctly (D25318752)

Sematic validator then recompute the strictness again and
verify/update the ESTree Node!
1. for non-lazy fn, it scan its body node to set FC
2. for lazy fn, it reuses the node strictness set by lazyParse to
set FC (D25318752)

### Proposed Changes

Parser is no longer responsible for setting strictness to node.
It needs to know strictness only for the parsing purpose.
1. preParse collects seen directives for every function
2. lazyParse reconstruct the directive nodes in lazy function body

Semantic validator is the source of truth for compute strictness
(and any other directive information) going forward.

(Note: this ignores all push blocking failures!)

Reviewed By: avp

Differential Revision: D29035883

fbshipit-source-id: fb79e5d1750c338cd3d3d7dcbf4659c24edf87e8