Preserve non-default function source.
Summary:
This diff implement the notion of "source code representation",
which can either be the actual source code, or a imitation of
NativeFunction for the purpose of implementation-hiding depending
on the source visibility of the function.

During HBC, functions are traversed and non-default source can
be preserved (interned) into the string table, allowing us to
retrieve them back in the runtime. Adding a
'function <name>() { [native code] }' for every of impl-hidden
functions would be a huge waste of space, so instead we associate
all of them with an empty string to 'mark' them and make this
this feature "almost free". See comments in the `Function::
getSourceRepresentationStr` for more details.

This is the next step to implement the Stage2 proposal "Function
Implementation Hiding" and the 'show source' asked by OSS.
- https://github.com/tc39/proposal-function-implementation-hiding

Reviewed By: tmikov

Differential Revision: D26468112

fbshipit-source-id: db5e1ca64f4243aac44ee51e41e4cabbf7182711