Encode native pointers as doubles
Summary:
We currently store native pointers under the NativeValue tag, which
corresponds to either a native integer, pointer, or just some bits.
This gives us 48 bits to store the native pointer, since it has to be
NaN-boxed like any other value.

New versions of Android have added use of a hardware assisted memory
safety feature called MTE. When enabled, it adds a tag in the 56th -
59th bits of heap allocations, and this tag can then be used to verify
that memory accesses use valid addresses. We can't NaN-box this
pointers because they no longer fit in 48 bits.

Conveniently though, since the tag is guaranteed to always be in the
top byte, we can guarantee that when interpreted as doubles, these
pointers won't be valid NaNs. Furthermore, we happen to already know
all of the places where native pointers are stored in `HermesValue`s,
since the `NativeValue` tag is overloaded with multiple uses.

So for now, we can simply store native pointers in `HermesValue`s
bit-for-bit. The caller is responsible for keeping track of where the
native pointers are, and all other code (like the GC) will treat it as
a double and ignore it.

In the longer term, we should ideally remove native pointers altogether.

Reviewed By: kodafb

Differential Revision: D29866451

fbshipit-source-id: e6f81f46ea7a4df0751cdf9da1fb396b8dcda97f