Move WriteUtf8String functionality to stream_base.cpp
Summary:
Moved WriteUtf8String functionality to stream_base.cpp
because the same write functionality is used for both
tty and pipe. This move was primarily made to mirror
the file structure that node uses.

Additionally, since uv_write is async, the data string
and the uv_write_t object need to survive past the
end of the function. Thus we created a callback where
these allocated objects are freed, and added the uv_run
line to ensure that the callback is called.

Reviewed By: avp

Differential Revision: D29909965

fbshipit-source-id: e6b0b469eca26671d3051d201e85587c454ba767