Remove special case for numeric properties in StoreNewOwnPropertyInst
Summary:
Allow `StoreNewOwnPropertyInst` to have numeric properties. This
simplifies some of the lowering passes, and avoids us having to deal
with restrictions on the bytecode format until instruction selection.

With this, we will have the invariant that any time a property is
stored for the first time in an object literal, it will be done using
`StoreNewOwnPropertyInst`, which makes it much easier to reason about
property ordering.

Reviewed By: tmikov

Differential Revision: D29377290

fbshipit-source-id: 52fd4cad6cacee2ce53cb4a2c3f68a2f400a05bd