Only consider StoreNewOwnPropertyInst in LowerAllocObject
Summary:
We currently don't distinguish between `StoreNewOwnPropertyInst` and
`StoreOwnPropertyInst` in `LowerAllocObject`. In theory, this could
mean that if we have both instructions storing to the same property, we
could end up with both versions being included in the buffer
instruction created by the pass.

In practice, this doesn't happen for string properties because
`LoadConstants` is run before this pass, and so string properties are
no longer `LiteralString` by this point. For numerical properties
however, this can cause problems.

Right now, the only way to emit a placeholder is to have an accessor
(although that will change in subsequent diffs). So if we run something
like:

```
var bar = {get 10 (){ return 42; }, 10 : 1, 11: 3}
print(bar[10])
```
which generates this IR:

```
neildhar@neildhar-mbp ~/build_hades_dbg> ~/.jsvu/hermes -dump-ir ~/tst.js
function global()
frame = [], globals = [bar]
%BB0:
  %0 = AllocObjectInst 2 : number, empty
  %1 = StoreNewOwnPropertyInst undefined : undefined, %0 : object, "10" : string, true : boolean
  %2 = StoreOwnPropertyInst 1 : number, %0 : object, "10" : string, true : boolean
  %3 = StoreNewOwnPropertyInst 3 : number, %0 : object, "11" : string, true : boolean
  %4 = StorePropertyInst %0 : object, globalObject : object, "bar" : string
  %5 = TryLoadGlobalPropertyInst globalObject : object, "print" : string
  %6 = LoadPropertyInst globalObject : object, "bar" : string
  %7 = LoadPropertyInst %6, 10 : number
  %8 = CallInst %5, undefined : undefined, %7
  %9 = ReturnInst %8
function_end
```
we will incorrectly consider both the placeholder emitted by the getter
and the value of `10`. This gets particularly bad because we allow
numeric properties to be reordered, and because `undefined` cannot be
in the buffer. So we end up creating a buffer for `{10 : 1, 11: 3 }`
and then having another instruction afterwards to store the `undefined`
placeholder value. This means that the above generates this bytecode:

```
Function<global>(1 params, 12 registers, 0 symbols):
Offset in debug table: source 0x0000, lexical 0x0000
    DeclareGlobalVar  "bar"
    NewObjectWithBuffer r1, 2, 2, 0, 0
    LoadConstUndefined r2
    PutOwnByIndex     r1, r2, 10
    GetGlobalObject   r0
    PutById           r0, r1, 1, "bar"
    TryGetById        r1, r0, 1, "print"
    GetByIdShort      r3, r0, 2, "bar"
    LoadConstUInt8    r0, 10
    GetByVal          r0, r3, r0
    Call2             r0, r1, r2, r0
    Ret               r0
```

and instead of printing `1`, we print `undefined`.

Reviewed By: tmikov

Differential Revision: D29377292

fbshipit-source-id: b92ddf14e61f96afbaab0caceac597f8d88694c0