Distinguish placeholders from full IR generation in object literals
Summary:
Create a new enum that tracks whether we have generated the full IR for
a node or only generated a placeholder in order to maintain insertion
order. This makes it clearer when we have inserted a placeholder and
should only be using `StoreOwn` as opposed to `StoreNewOwn`.

More importantly, this is necessary for the next diff which starts
adding placeholders for normal values as well.

Reviewed By: tmikov

Differential Revision: D29222538

fbshipit-source-id: 293191925f4c44919376a911074bb81e34a0d442