Insert placeholders for value expressions in object literals
Summary:
When we have multiple values with the same property name in an object
literal, we do not insert a placeholder value the first time a property
is inserted. This ends up meaning that we do not maintain the order of
properties correctly. For example, for the following code:
```
var foo = { a: 1, b: 2, a: 3 };
print(Object.getOwnPropertyNames(foo));
```

Hermes will print `b,a`, when it should be `a,b`.

Other engines do this correctly:
```
neildhar@neildhar-mbp ~> ./.jsvu/v8 tst.js
a,b
neildhar@neildhar-mbp ~> ./.jsvu/sm tst.js
a,b
neildhar@neildhar-mbp ~> ./.jsvu/javascriptcore tst.js
a,b
neildhar@neildhar-mbp ~> ./.jsvu/chakra tst.js
a,b
```

This diff adds a placeholder whenever we encounter an occurrence
of a property that is not the final value the property will have.

Reviewed By: tmikov

Differential Revision: D29222539

fbshipit-source-id: 09dafdf90a53a9c9c83eb8560d92cd41f57371cf