Some NodeLabel->StringLiteral, Vec<NodePtr>
Summary:
Changed some types from NodeLabel to StringLiteral. This broke some of
the printing and debugging code. Disabled it for now with
unimplemented!().

Changed NodeList to be Vec<Box<Node>> instead of Vec<Node>.

Changed the comment for SourceRange to indicate that the end location is
inclusive since it is hard to define exclusive location with
line/column.

Reviewed By: avp

Differential Revision: D30124002

fbshipit-source-id: 334ed819ca27e9b69928f3e9f81a35fe5ff42b54