Console table new functionality
Summary:
Maps, Sets, MapIterators, SetIterators are currently not
supported because they have long dependency chains
on functionalities that are not implemented by node-hermes
yet. Therefore the Node code is modified slightly to print to
stderr whenever an object of one of those types is passed in.
To be able to do this type detection, I also manually added
the type functions to types.js.
The only other major modification here was to get StringWidth
working,

Reviewed By: avp

Differential Revision: D30139089

fbshipit-source-id: 995c0d127ec18fc39c0572b330859dd48f83bc1b