A code generator for the Hermes parser FFI
Summary:
Rustgen is a C++ tool that generates a Rust source file for interfacing
with the C++ Hermes parser. The file is printed to stdout and contains
the following parts:

- NodeKind enum.
- AST attribute getters.
- A function recursively converting each Hermes AST node to a Juno AST
  node.

To generate the files (in future diffs):
```
cd juno/src/hparser
rustgen generated_ffi.rs
cargo fmt
```

We haven't found a reliable way to exclude the generated file from
`rustfmt`, so we might just as well format it manually when generating.

Reviewed By: avp

Differential Revision: D30180410

fbshipit-source-id: 5c3812b07e3cf1bb4a1ba4630e7dd8a15221711a