Increase the default max heap size from 512MiB to 3GiB
Summary:
The default max heap size GC config of Hermes, 512MiB, is too small to
be the default config for React Native apps from the nowadays standard.

It casued memory issues on Android (#295) and was mediated by manually
[overriding this number to 1024MiB from the RN Android code path](https://github.com/facebook/react-native/commit/63d20d3b1ef35cb4398d63d62f631f7f5d2935c7),
and a similar OOM issue was reported on iOS recently (#551).

As Hermes being adopted by more platforms (Windows, macOS), it make no
sense to override this number again and again in each platforms.
Rather, the default max heap should be changed once and for all,
The RN side of change is tracked at https://github.com/facebook/react-native/pull/31900.

We choose 3GiB which is "unlimited" in practice, making Hermes closer
to JSC which doesn't have such limit as well.

Reviewed By: neildhar

Differential Revision: D30181817

fbshipit-source-id: 54a442e43b7884a9f5d7ed099cb9b9866ee4f109