Make enums for AST properties
Summary:
Use a new macro (in `support` for now) called `define_str_enum`
which creates an enum that easily goes back and forth from strings
to restrict the types on the AST's `operation` properties.

Use of generics allows us to also cut down the amount of code required
in the `hparser::convert` module by writing a single `cvt_enum` function
that can be called in the appropriate place in `rustgen`.

Implement the traits required for AST children and fix any remaining code.

Reviewed By: tmikov

Differential Revision: D30400454

fbshipit-source-id: b626eded05f0c46df8c85c29a977b0e913728819