Add CMake lean build
Summary:
Add the ability to build libhermes.so with a lean VM in CMake. This
reduces the size of the Hermes binary by ~25%.

Aside from libhermes itself, the distinction is made by having separate
targets, which allows tests to continue running correctly. However, in
order to keep the library name correct, we have to toggle the build
type of libhermes using a build flag.

Reviewed By: kodafb

Differential Revision: D30329792

fbshipit-source-id: c51f5b02d132993a04070a09d19d67c990c26245