improve scanning of magic block comments
Summary:
Magic block comments did not go through the normal comment scanning
process, leading to a couple of subtle problems:
- Unicode line terminator sequences were not recognized as new line.
- The magic comment was not checked for invalid UTF-8 sequences.

This change passes everything through the same block comment scanning
code which checks for Unicode line terminators and validates UTF-8
characters.

This means that now there is a guarantee that if there are no source
errors, the stored comments and the magic comments are guaranteed to be
valid UTF-8. If there is at least one error, they may contain invalid
UTF-8.

Reviewed By: avp

Differential Revision: D30388860

fbshipit-source-id: e0ec279185c71d37b48448e0e85adf2c93185ff1