introduce NullTerminatedBuf
Summary:
The various Hermes parsers require that their input be null terminated.
That is an invariant that has proven difficult to maintain in practice
because it is not explicit in the C++ interfaces.

`NullTerminatedBuf` aims to solve this problem when Hermes parsers are
exposed to Rust by making the invariant explicit. It provides a
reference to a slice that is always explicitly null terminated.
Internally it either just borrows a reference, or owns a copy of the
original non-terminated input.

`NullTerminatedBuf` can be created from `&[u8]`, `&str` and by reading
from a file. The first two provide options to check whether the input is
already null-terminated and just borrow it instead of copying. That can
be useful if the input is already null terminated for another reason.

The more interesting from performance perspective is creating a
`NullTerminatedBuf` from a file. In terms of functionality it is
equivalent to `llvm::MemoryBuffer::getFile()`.

The current implementation is extremely primitive, it simply reads the
entire file into a Vec and appends a 0. There are many possibilities for
optimizations, but applying them was not trivial since we weren't able
to find portable ways for checking if a file is a "regular file" (as
opposed to a pipe, a character device, etc) in Rust. Listing them here
for posterity:

- Obtain the size the file in advance to avoid reallocations and to
  minimize number of syscalls (hopefully down to one, if not
  interrupted).
- If the file is not aligned to page, memory map it. The "filler space"
  past the end is guaranteed to be zeroes.
- Expose `llvh::MemoryBuffer` through a couple of simple C wrappers and
  just use that.

Reviewed By: avp

Differential Revision: D30444136

fbshipit-source-id: 9b1d198a6de385ed075d89a1229313e14c56ad0a