Move Android build flags into gradle file (#579)
Summary:
Pull Request resolved: https://github.com/facebook/hermes/pull/579

Overriding `CMAKE_CXX_*_FLAGS` for a specific build mode is error
prone. For instance, it excludes .c files. Instead, we can specify the
build mode in the gradle file. This should be a slight improvement in
dev mode performance, and also slightly shrinks the size of a release
libhermes.so.

While here, also simplify the other flags being set for Android by
removing flags that do not seem to affect the build.

Reviewed By: kodafb

Differential Revision: D30463984

fbshipit-source-id: 76ce554d908a2547836dbb97ce00a4b92a646aaa