Separate key and value processing for objects in buffer
Summary:
We currently have two completely independent codepaths for
constructing object literals depending on whether we get a cache hit.
Instead, we can simplify the code by factoring out the logic to check
and populate the cache. This lets us construct the hidden class in
one pass, and then efficiently fill the property slots in a newly
created object.

While this slightly improves performance on its own, the end goal is
to disentangle the iteration over the two serialised buffers, so that
we can optimise them individually.

Reviewed By: kodafb

Differential Revision: D29727159

fbshipit-source-id: 8371b105038dbb0be5309e42d21ef78ca6851629