Store offsets in contiguous array
Summary:
Instead of storing each the metadata for each field type in its own
heap allocated array, we can reduce the number of indirections and
improve locality by storing all the offset information inline.

This diff lays out all of the offsets in a single contiguous array that
the marking loop can then iterate over. Instead of having separate
arrays entirely, we just need to keep track of where a given slot type
begins and ends inside this contiguous storage.

This means that instead of following an indirection for each field
type, the marking loop can quickly index into the metadata table given
a CellKind.

The offset and array information are stored adjacent to each other
since they will be used together most of the time.

While we're here, this diff also shrinks the ArrayType tag to be a
single byte.

A possible follow up would be to move the VTable* and names out of
the Metadata struct. These pointers are huge relative to the rest of
the metadata and moving them out will make all the metadata for a
given type fit in ~20 bytes.

Reviewed By: kodafb

Differential Revision: D29448268

fbshipit-source-id: 597930732c3aac52f7545fb828d9e34901141159