Use preprocessor to handle slot types
Summary:
We currently have a lot of boilerplate to handle each of the possible
slot types. Now that the earlier diff also introduces dependencies on
the order in which we traverse the offsets, it makes sense to use the
preprocessor to reduce the amount of boilerplate and simplify handling
the order of slot types.

Reviewed By: kodafb

Differential Revision: D29525615

fbshipit-source-id: 3aa13dbb2582478f574e40bb9d3c8b1b4b2c50f1