Remove custom VTable from TypedArray
Summary:
Since there are only two functions that require VTable based dispatch
in TypedArray, and since we have an easy way of enumerating all the
types in a single translation unit, we don't actually need to use a
VTable. We can instead keep an array of functions in
`JSTypedArrayBase::allocate` and `JSTypedArrayBase::allocateSpecies`.

Reviewed By: kodafb

Differential Revision: D29541676

fbshipit-source-id: ce40d313441bd0a5145aff25a857ac97746cc28c