Speed up location finding by caching cur line
Summary:
If locations are accessed in a natural left to right order, they will
always be either in the current or the next line, which we can cache to
avoid almost 100% of the lookups.

That is more the case for Juno than for Hermes, but it should benefit
Hermes too.

Juno needs access to the whole line anyway, because it needs to operate
with Unicode character indices, not byte offsets, so this kind of
caching integrates naturally into that model.

Reviewed By: avp

Differential Revision: D30720707

fbshipit-source-id: 5fc68250fdf7d758d5a79ac0a012734a0bed8d1c