utility to fetch data from an URL
Summary:
fetch_url fetches all data from an URL and returns it. For now only
file: and data: URLs are supported, but more (like http:) can be added.

It returns the entire data instead of a reader for simplicity. Juno is
not a multithreaded server so it doesn't need to deal with reading a
file incrementally.

Reviewed By: avp

Differential Revision: D31044163

fbshipit-source-id: 1b3f4a3ee75744305019d6c4a8a4b569fa3fa473