Reverse stack direction
Summary:
In a downward growing stack, increasing register offsets are found at
lower addresses. This means that when we load from a register, we are
applying a negative offset to some pointer.

In arm64 and x86, negative offsets are not part of the addressing mode
in load instructions. This means that instead of being able to apply
the offset and load in a single instruction, we first need an
instruction to perform the subtraction and compute the address.

See https://godbolt.org/z/qGW4a8d5a

This diff reverses the direction of the stack completely, so that it
grows upwards instead, allowing register loads to become more efficient.

Reviewed By: kodafb

Differential Revision: D30803535

fbshipit-source-id: 29010a9998a4b75dd68ebd45bb014e18820d1e84