Remove dead_code ignores
Summary:
Exported symbols from a library won't warn when unused.
Get rid of the annotations at the top level of modules so we get the
warnings back. We can use `#[allow(dead_code)]` at a more granular
level if necessary in the future, but as it is it isn't hiding any warnings.

Reviewed By: tmikov

Differential Revision: D31208946

fbshipit-source-id: e1745c1177b0ef3b387e51bc6dc0a61ee9c1e89a