Allow reading hermes_build from a TOML file
Summary:
Read `.hermes.toml` in the `juno` root if the env variable
doesn't exist.

Currently there is some code duplication, so that will have to be
encapsulated perhaps in a separate crate when we start building
with CMake. This is just a simple solution for now until we do
the CMake build properly.

Reviewed By: tmikov

Differential Revision: D31324689

fbshipit-source-id: 4355a9b8591fd971efdbdbd48013a63a84247422