Add sourcemap output from CLI
Summary:
Allow the user to output the full sourcemap to a `.map` file.
To do this, also allow the user to output the actual output to whatever
file they would like.
To work around limitations of `clap`, add an `Opt::validate` function
that lets us do whatever post-validation we want with flexibility.

Reviewed By: tmikov

Differential Revision: D31347358

fbshipit-source-id: 7c6f3b2d0dc8dba4064761518e9b8c35039f6209