Simple SourceManager
Summary:
SourceManager owns a collection of source buffers with names (so they
can be parsed and so it can provide context when printing errors).

This is a very simple implementation - it directly prints error messages
to stderr without any buffering, etc.

The error() method takes `&self` by design - the implementation will use
interior mutability. A SourceManager is intended to be easily shareable.

Reviewed By: avp

Differential Revision: D31311326

fbshipit-source-id: 47d48d2efb3bcf2978d73b524cfd7e8b4122a8be