Enforce priority of mutator
Summary:
We currently transfer control between the background thread and mutator
by periodically unlocking `gcMutex_` on the background thread to give
the mutator a chance to acquire it. This seems to work, but it leaves
the scheduling decision up to the OS, which may choose to never yield
to the mutator.

However, we actually want the background thread to always yield to the
mutator, to avoid blocking JS execution. We can do this using a flag
and a condition variable to ensure that the mutex is released and
acquired by the mutator.

Reviewed By: tmikov

Differential Revision: D31323578

fbshipit-source-id: f9f0c5bf41fca8ab8585d2913121905d7b2bdfb5