Only acquire weak ref mutex when needed
Summary:
We don't need to hold the weak ref mutex the entire time we are
marking. Instead, we can just acquire it whenever we are actually about
to invoke a weak ref callback on some object, which should be fairly
rare.

This reduces contention for the mutex, which should improve performance.

Reviewed By: kodafb

Differential Revision: D31323956

fbshipit-source-id: 7118fc29fc0b9805cd8c0ae66f0748b171fa7964