Create separate functions for DataView get and set
Summary:
Some compilers (like clang on Windows) can deduplicate the code
generated by a template, causing two separate functions to map to the
same address. This happens with some of these `DataView` functions for
instance, which can be templated on both `uint8_t` and `int8_t`. In
theory, this could be a useful optimisation, but given that other
compilers don't do it, and that it interferes with our ability to get
the function name string from a function pointer, we can avoid it for
now.

This diff creates multiple separate non-template functions for each
`TypedArray` type instead. Combined with the next diff, it also allows
us to delete the `NATIVE_FUNCTION_TYPED` macro, which simplifies our
handling of native functions in the preprocessor.

Reviewed By: avp

Differential Revision: D31781921

fbshipit-source-id: 898bd552dd597a077045f36434a0369e2ca75dff