Create separate functions for TypedArray constructors
Summary:
Move the existing templates to generate TypedArrray constructor code
into an anonymous namespace, and instead generate separate standalone
functions for each one. Similar to the previous diff, this avoids the
compiler trying to deduplicate code, and saves us needing to deal with
templates in the places where we handle native functions.

Reviewed By: avp

Differential Revision: D31781920

fbshipit-source-id: e1dd082236935b47960a7ee31885adb9a9c654ac