improve Debug printing of Atom-s
Summary:
Printing Atom is not very informative, because they are just numbers
(unlike Hermes where they also point to the actual string).

Add mechanisms to optionally register an AtomTable as "active" in the
current thread. Implement a custom Debug formatter for Atom, which
checks if there is an active table and the atom is valid inside it, and
prints the corresponding string.

The safe mechanism is a function that sets the active table, invokes a
callback, and restores the active table.

Also provide an unsafe mechanism to just set the table, in case urgent
debugging is needed. It is unsafe, because there is no guarantee that
the table would not be moved or destroyed.

Reviewed By: avp

Differential Revision: D31878793

fbshipit-source-id: 49db11e414610643e0fd137e7938e5d53ca9c582