Fix test for double truncation
Summary:
Work around Clang 13 treating the truncation fast path as UB and
optimising it to not work on compile time constants.

This appears to be a bug in clang, since adding the
`-fno-strict-float-cast-overflow` flag does not fix it. Nonetheless,
this diff also adds that flag, since it is generally desirable for
Hermes code.

Reviewed By: tmikov

Differential Revision: D31939409

fbshipit-source-id: b86830d129e6bf1b872e502294ab9ea1893ce242