validator returns std::error::Error
Summary:
It is more convenient if the AST validator uses the source manager to
report the errors and returns a std::error::Error, which can easily be
reported automatically.

Rename the existing function to `validate_tree_pure()` (since it doesn't
modify any state), and implement a simple wrapper `validate_tree()`,
reporting to SourceManager and returning Error.

Reviewed By: avp

Differential Revision: D32012703

fbshipit-source-id: 1401ed1ac99e7fd41b5adfba94dbc79b83bf4ccb