Avoid emitting extra semicolons for certain statements
Summary:
`if` and related statements were emitting extraneous `;`.
For example, if we have the AST for:
```js
if (x)
  y()
else
  z()
```
then we should only emit `;` for the `y()` and `z()` statements,
but not for the `if` or either the consequent or alternate.

To do this, change `ends_with_block` (which was only used in this one
place) and make it `stmt_skip_semi`, which indicates whether it's OK
to skip the semicolon on a given statement node. On `if` statements,
we must skip the semicolon because the semicolon will be emitted
when emitting the consequent or alternate. Same goes for `for`, `while`,
etc.

Reviewed By: tmikov

Differential Revision: D31937599

fbshipit-source-id: 1004258fee50e326a537765875782014f45b77ac