Create PassManager and run optimization passes
Summary:
Add a new crate `pass` for writing and managing optimization passes
over Juno representations.
It exposes `PassManager` and `Pass`, which allow for users to create
and execute optimization pipelines.

To demonstrate that it works, create an `AddNegative` pass
and add it to the pipeline the `juno` CLI uses when given `-O`.

Reviewed By: tmikov

Differential Revision: D31998123

fbshipit-source-id: ffcb0d959b880c318e4f5269406f7b19bd2d6b74