Emit extra decimal point in numeric member expressions
Summary:
To handle cases like `50..toString()` we need to ensure that the
number has a decimal point in it.

The other way to handle this would be to pass a flag through to the C++
function that's doing the work, but given that number literal strings
are typically small and this is rare, special casing in `gen_js` seemed
easier.

Reviewed By: tmikov

Differential Revision: D31967210

fbshipit-source-id: 191c94ded5732e5a5be749ae6f851c9b2a3325f3