Use a macro for RuntimeGC kinds
Summary:
Use a macro so that if we ever use the RuntimeGC mechanism again in
the future, we don't accidentally miss places to update. This risk is
particularly high because it is difficult to make sure multiple
available GCs are fully tested, particularly when dealing with things
like minimum and maximum allocation sizes, which may only trigger bugs
under certain configurations (like without compressed pointers).

Reviewed By: kodafb

Differential Revision: D31642786

fbshipit-source-id: acbfe46480d7d6f3a7c6e3d3ed300fafa927c861