Basic dynamic threshold for concurrent collections
Summary:
Hades currently uses a fixed 75% threshold for starting full
collections. This is very conservative, and results in a large number
of unnecessary collections.

This diff changes the collection threshold so that it is computed
dynamically for concurrent mode, which allows us to dramatically reduce
the number of full collections. However, it doesn't make any change to
incremental mode, since increasing the threshold there needs to be
balanced with increased pause times, and we're not currently set up to
do that easily.

Reviewed By: kodafb

Differential Revision: D31365877

fbshipit-source-id: d6c6dceeb88785e48f8902bbe044b3089d7a918d