Add buffer before compacting in large heaps
Summary:
In large heaps, it is likely that variations in the heap size from one
collection to the next can be very large in terms of MB. However, the
compaction code as written doesn't deal with this well. Currently,
we'll trigger a compaction if the target size drops by more than 1
segment. In small heaps, this is not a concern since it represents a
4MB change.

However, once the heap size is in the hundreds of MB, even with the
moving average added in the previous diff, it is likely that the target
size will vary by more than one segment from one collection to the
next. In that case, it is useful to have a slightly more conservative
compaction criteria, that scales with the size of the heap.

This diff makes it so that the actual size can be up to 5% larger
than the target size.

Reviewed By: kodafb

Differential Revision: D32004981

fbshipit-source-id: 6e8cd26a472fae5aca236a2d91384820b126733d