new module decl_collector.
Summary:
`decl_collector` collects all declarations in a function and associates
them with scopes. This is necessary because all declarations need to be
hoisted either to the top of the function or the top of their own scope.

This is split for an easier review. It is used by a later diff.

Reviewed By: avp

Differential Revision: D32013211

fbshipit-source-id: 074296f36847fb18896ee355d3495bed2b9390b4