Semantic resolution
Summary:
The resolver returns a populated SemContext. The ES5 features should be
complete, many ES6 features are lacking.

It should be considered WIP, however barring any critical bugs, it
should be able to land and iteration can happen in following diffs.

There is a profound lack of tests, partly because it is not entirely
clear what the best way of testing would be. Error messages are not
sufficient, perhaps some form of structured printing of all data (that
is what the C++ version did, but it was never used).

Reviewed By: avp

Differential Revision: D32040069

fbshipit-source-id: 2609e3bdbc791004c8ea30b12a27f17fd701fb4f