support comments in mutations
Summary:
Using the prettier code forked in the previous diff we can attach comments to the relevant nodes. Because of how the infra was setup - this means almost no work was required to ensure comments are kept.

As we discussed - I added an option (default false) to replacements which allow you to keep the comments when the node is replaced.

Reviewed By: pieterv

Differential Revision: D32301464

fbshipit-source-id: fb729464684751e5568ab4b505c3276857ccd94b