Move LIT tests to cli crate
Summary:
Keeping the LIT tests in the `juno` crate meant that there was effectively
an unrecorded dependency on the `cli` crate, because that's where
the binary we use to run the LIT tests comes from.

Moving the LIT test runner to the `cli` crate correctly uses the binary
within that crate, which will be correctly rebuilt by `cargo test`.

Reviewed By: tmikov

Differential Revision: D32442240

fbshipit-source-id: 227c039fffe5d19cf251c63a8086eab9258e3e96