compile arbitrarily deep +,= and assignment chains
Summary:
Since Hermes parses, validates and generates IR recursively, it has
pretty strict limit on the nesting level of AST. This can be a problem
when comparing to other engines, which have higher limits. We fail
sooner when compiling synthetic expressions like '0 + 0 + 0 ...' or
`a = a = a = a = ...`.

It can also be an actual problem, demonstrated by a recent bug report
about 1800 nested assignments:

https://github.com/facebook/hermes/issues/628

Fortunately, there are ways to fix this. We already parse binary
expressions non-recursively, so the parser imposes no limit on their
nesting. In the previous commit we changed parsing of assignment
expressions to be also non-recursive, while setting am artificial limit
of 30,000 nesting level.

In this commit we take this one step further and remove the limit in
semantic validation and IRGen. The solution is based on two new generic
methods, which turn nested expressions of a certain kind into a list of
arbitrary length. Then we can simply process the list non-recursively.

This commit doesn't apply this technique everywhere - only for "+/-" and
ordinary assignment. However it will be trivial to extend to other
operators and assignment kinds.

Closes #628

Reviewed By: avp

Differential Revision: D32486320

fbshipit-source-id: a2ed392ad9ed15324cef8a0f3c931546710f2e8a