parser nested assignments non-recursively
Summary:
Parse assignments of the form `a = b = c = ...` non-recursively using an
explicit stack. This is the first step to lifting the limits on number
of nested assignments.

Note that the resulting AST still cannot be dumped, because the dumping
is recursive. However that is a bug for another time.

Reviewed By: avp

Differential Revision: D32486322

fbshipit-source-id: e2b66aaaa1a75be871ecaec1c714c33daff90cfe