Efficiently set existing CompressedPointer in GCPointerBase::set
Summary:
Save a compressed pointer encode when setting a `GCPointer` to another
`GCPointer`. We still have to decode it for the write barrier, although
it's possible the compiler can optimise it by moving the decode into
the write barrier slow path for us. Either way, encodes are generally
more expensive than decodes.

Reviewed By: kodafb

Differential Revision: D32863388

fbshipit-source-id: 8a41ee5b949678f7848c455af28cf3cf688e6cb8