Forbid nullptr HermesValues on the heap
Summary:
Disallow nullptr `HermesValue`s and `HermesValue32`s on the heap. This
allows us to use the more efficient non-null operations on compressed
pointers. Currently, this is only an improvement in the decode path,
however, in a non-aligned contiguous heap, this will also be an
optimisation for the encode path. In addition, as seen in the previous
diff, it also allows us to optimise write barriers for the non-null case.

This has some drawbacks:
1. It hurts convertibility of any HV to HV32, since `nullptr` HVs can
no longer be encoded in HV32.
2. It adds some overhead and risk since we need to worry about whether
a given HV may be a nullptr.

However, I convinced myself that these are acceptable because of the
performance improvements, and the following considerations:
1. We already cannot convert between them perfectly, since things like
native ints and pointers have no representation in HV32. This just adds
another restriction that we need to be aware of.
2. Only object pointers in HV32 are currently nullable. String pointers
are `vmcast` to `StringPrimitive`, so they cannot be null, and boxed
doubles can never be null either.
3. There are many places already where nullptr is not an acceptable HV,
since it may be visible to JS. We already have to be careful in such
cases.

Reviewed By: kodafb

Differential Revision: D32873122

fbshipit-source-id: 71a3339825b0d3361caadff89ecfc575855496ba